chore: move index and export everything to clean up import paths of dependencies
